Fix FC in Bluetooth due to MAP
This commit fixes a FC in the MAP implementation when a vehicle sends multiple MNS notification requests. The old implementation would disconnect and then reconnect, however this seems to cause timeout problems on some vehicles. This specific behaviour is not specified, and there seems to be nothing wrong with re-using the existing obex connection.

Change-Id: I165544449180334e35c1f9510ace8d11ee215a55